+++ include/oglplus/images/newton.hpp
@@ -63,7 +63,7 @@ private:
 		return a*(1.0f - coef) + b*coef;
 	}
 
-	template <typename Function, typename Mixer, size_t N>
+	template <typename Function, typename Mixer, unsigned N>
 	void _make(
 		GLsizei width,
 		GLsizei height,
@@ -85,7 +85,7 @@ private:
 				_mix(lb.x(), rt.x(), float(i)/float(width-1)),
 				_mix(lb.y(), rt.y(), float(j)/float(height-1))
 			);
-			size_t n, max = 256;
+			unsigned n, max = 256;
 			for(n = 0; n != max; ++n)
 			{
 				Vec2f zn = z - _cdiv(
