+++ include/oglplus/shapes/sphere.hpp
@@ -30,7 +30,7 @@ class Sphere
 {
 private:
 	GLdouble _radius;
-	size_t _sections, _rings;
+	unsigned _sections, _rings;
 public:
 	/// Creates a sphere with unit radius centered at the origin
 	Sphere(void)
@@ -40,7 +40,7 @@ public:
 	{ }
 
 	/// Creates a sphere with unit radius centered at the origin
-	Sphere(GLdouble radius, size_t sections, size_t rings)
+	Sphere(GLdouble radius, unsigned sections, unsigned rings)
 	 : _radius(radius)
 	 , _sections(sections)
 	 , _rings(rings)
@@ -57,17 +57,17 @@ public:
 	GLuint Normals(std::vector<T>& dest) const
 	{
 		dest.resize(((_rings + 2) * (_sections + 1)) * 3);
-		size_t k = 0;
+		unsigned k = 0;
 		//
 		GLdouble r_step = (1.0 * math::pi()) / GLdouble(_rings + 1);
 		GLdouble s_step = (2.0 * math::pi()) / GLdouble(_sections);
 
-		for(size_t r=0; r!=(_rings+2);++r)
+		for(unsigned r=0; r!=(_rings+2);++r)
 		{
 			GLdouble r_lat = std::cos(r*r_step);
 			GLdouble r_rad = std::sin(r*r_step);
 			// the sections
-			for(size_t s=0; s!=(_sections+1);++s)
+			for(unsigned s=0; s!=(_sections+1);++s)
 			{
 				dest[k++] = r_rad * std::cos(s * s_step);
 				dest[k++] = r_lat;
@@ -85,13 +85,13 @@ public:
 	GLuint Tangents(std::vector<T>& dest) const
 	{
 		dest.resize(((_rings + 2) * (_sections + 1)) * 3);
-		size_t k = 0;
+		unsigned k = 0;
 		//
 		GLdouble s_step = (2.0 * math::pi()) / GLdouble(_sections);
 
-		for(size_t r=0; r!=(_rings+2);++r)
+		for(unsigned r=0; r!=(_rings+2);++r)
 		{
-			for(size_t s=0; s!=(_sections+1);++s)
+			for(unsigned s=0; s!=(_sections+1);++s)
 			{
 				GLdouble vx = std::cos(s*s_step);
 				GLdouble vz = std::sin(s*s_step);
@@ -122,15 +122,15 @@ public:
 	GLuint TexCoordinates(std::vector<T>& dest) const
 	{
 		dest.resize(((_rings + 2) * (_sections + 1)) * 2);
-		size_t k = 0;
+		unsigned k = 0;
 		//
 		GLdouble r_step = 1.0 / GLdouble(_rings + 1);
 		GLdouble s_step = 1.0 / GLdouble(_sections);
-		for(size_t r=0; r!=(_rings+2);++r)
+		for(unsigned r=0; r!=(_rings+2);++r)
 		{
 			GLdouble r_lat = 1.0 - r*r_step;
 			// the sections
-			for(size_t s=0; s!=(_sections+1);++s)
+			for(unsigned s=0; s!=(_sections+1);++s)
 			{
 				dest[k++] = s * s_step;
 				dest[k++] = r_lat;
@@ -169,16 +169,16 @@ public:
 	{
 		assert((1<<(sizeof(GLushort)*8))-1>=((_rings+2)*(_sections+1)));
 		//
-		const size_t n =
+		const unsigned n =
 			2 * (_rings + 1)*(_sections + 1);
 		//
 		IndexArray indices(n);
-		size_t k = 0;
-		size_t offs = 0;
+		unsigned k = 0;
+		unsigned offs = 0;
 		// the triangle strips
-		for(size_t r=0; r!=(_rings+1); ++r)
+		for(unsigned r=0; r!=(_rings+1); ++r)
 		{
-			for(size_t s=0; s!=(_sections+1); ++s)
+			for(unsigned s=0; s!=(_sections+1); ++s)
 			{
 				indices[k++] = offs + s;
 				indices[k++] = offs + s + (_sections+1);
@@ -195,7 +195,7 @@ public:
 	DrawingInstructions Instructions(void) const
 	{
 		auto instructions = this->MakeInstructions();
-		for(size_t r=0; r!=(_rings+1); ++r)
+		for(unsigned r=0; r!=(_rings+1); ++r)
 		{
 			DrawOperation operation;
 			operation.method = DrawOperation::Method::DrawElements;
